project2/requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
tqdm
Pillow
opencv
scikit-learn
matplotlib
streamlit
fastapi
uvicorn
pillow